feat: self-hosting
