+++ package.json
@@ -6,6 +6,7 @@
   "description": "TypeScript Module Loader",
   "license": "MIT",
   "bin": "dist/runtime/bin.js",
+  "types": "dist/types/index.d.ts",
   "contributors": [
     {
       "name": "C. Lewis",
@@ -19,8 +20,8 @@
     }
   ],
   "exports": {
-    ".": "./dist/runtime/loader.js",
-    "./config": "./config/index.js",
+    ".": "./loader/index.js",
+    "./*": "./dist/*/index.js",
     "./package.json": "./package.json"
   },
   "files": [
@@ -34,7 +35,7 @@
     "node": ">=16"
   },
   "scripts": {
-    "bootstrap": "esbuild src/commands/build/index.ts --target=es2021 --outfile=dist/bootstrap.js && TSM_BOOTSTRAP=1 node dist/bootstrap.js",
+    "bootstrap": "esbuild src/commands/build.ts --target=es2021 --outfile=dist/bootstrap.js && TSM_BOOTSTRAP=1 node dist/bootstrap.js",
     "prebuild": "yarn bootstrap",
     "build": "./src/index.ts build",
     "types": "tsc",
@@ -47,12 +48,12 @@
     "fast-glob": "^3.2.10"
   },
   "devDependencies": {
-    "@types/node": "16.11.6",
-    "@types/react": "17.0.33",
+    "@types/node": "17.0.8",
+    "@types/react": "17.0.38",
     "@typescript-eslint/eslint-plugin": "^5.9.1",
     "@typescript-eslint/parser": "^5.9.1",
     "eslint": "^8.6.0",
-    "typescript": "4.4.4"
+    "typescript": "4.5.4"
   },
   "keywords": [
     "esm",
