+++ src/commands/build.ts
